feat: 반응형 레이아웃 css 미디어 쿼리 방식으로 변경
